requirements.txt
fastapi
requests
numpy
python-dotenv
flask
